GameTaskDraw + MenuButton: fix drawDt over-accumulation + float->u16 wasm trap (#347)
Two faithful fixes for wasm-only crashes on screen transitions, found via a new
UBSan repro (test_mainscreen_to_gamemode):

- GameTaskState.cpp GameTaskDraw @0x00119dfc: the port gated BOTH game_work.dt=
  s_drawDt AND the s_drawDt=0 reset inside the draw-gate; the binary does both
  UNCONDITIONALLY (only the draw dispatch is gated). During a transition the gate
  is off (s_updated cleared by Exit), so s_drawDt accumulated across frames -> a
  huge dt on the first post-transition draw. Now unconditional, matching binary.

- MenuButton.cpp:896: SinIdx((uint16_t)(m_NewIndicatorTimer*180*182)) cast a
  float straight to unsigned short; the huge dt pushed the product to 65694.7 >
  65535. Direct out-of-range float->u16 is silent UB on x86 but TRAPS on wasm
  (trunc_f32_u). Binary MenuButton::Draw @0x0019c2e4 does (int)(...)&0xffff (wrap
  mod 65536); ported as float->u32->u16 (bit-exact wrap, in-range intermediate).

UBSan-confirmed: MenuButton:896 error gone after the fix. (Note: this is NOT the
GameModeScreen::CreateControls crash from the user's symbolized trace -- that's a
separate m_RingTex/texture-vtable issue being chased next.)

+++ src/hud/MenuButton.cpp
@@ -893,7 +893,11 @@ void MenuButton::Draw(float* hudScaleRaw) {
         IngamePopup* popup = GetIngamePopup(0x10);
         if (popup) {
             float scale = (m_RestScale.x != 0.0f) ? (size.x / m_RestScale.x) : 0.0f;
-            const float sinV = SinIdx((uint16_t)(m_NewIndicatorTimer * 180.0f * 182.0f));
+            // v1.6.1 MenuButton::Draw @0x0019c2e4: SinIdx angle index wraps mod 65536
+            // (binary: `(int)(timer*180.0*182.0) & 0xffff`). float->u32->u16 replicates that
+            // wrap; a direct float->u16 cast is UB on native and TRAPS on wasm when the
+            // product exceeds 65535 (#346).
+            const float sinV = SinIdx((uint16_t)(uint32_t)(m_NewIndicatorTimer * 180.0f * 182.0f));
             const float bob  = (sinV < 0.0f ? -sinV : sinV) * 8.0f;
             Vec3 anchor = GetAdjustedPos();
             anchor.x += m_ShakeScale.y * m_RestScale.x * 0.5f;
